mod refactoring
