+++ kwap/src/std/net/convert.rs
@@ -1,128 +1,20 @@
-use core::convert::Infallible;
-use std::collections::HashMap;
-use std::io;
-use std::net::UdpSocket;
-use std::ops::Deref;
-use std::sync::{Arc, Mutex};
+use ::std::io;
 
-use kwap_common::prelude::*;
-
-use crate::net::{Addrd, Socket};
-use crate::todo::ResultExt2;
-
-/// TODO
-#[derive(Debug, Clone)]
-pub struct UdpStream {
-  sock: Arc<UdpSocket>,
-  addr: no_std_net::SocketAddr,
-  tx_buf: Vec<u8>,
-}
-
-impl UdpStream {
-  fn new(sock: Arc<UdpSocket>, addr: no_std_net::SocketAddr) -> Self {
-    Self { sock,
-           addr,
-           tx_buf: vec![] }
+pub(crate) fn io_to_nb(err: io::Error) -> nb::Error<io::Error> {
+  match err.kind() {
+    | io::ErrorKind::WouldBlock => nb::Error::WouldBlock,
+    | _ => nb::Error::Other(err),
   }
 }
 
-impl io::Write for UdpStream {
-  fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-    self.tx_buf = [&self.tx_buf, buf].concat();
-    Ok(buf.len())
-  }
-
-  fn flush(&mut self) -> io::Result<()> {
-    Socket::send(self.sock.as_ref(), Addrd(&self.tx_buf, self.addr)).map_err(nb_to_io)
+pub(crate) fn nb_to_io(err: nb::Error<io::Error>) -> io::Error {
+  match err {
+    | nb::Error::WouldBlock => io::Error::from(io::ErrorKind::WouldBlock),
+    | nb::Error::Other(err) => err,
   }
 }
 
-impl io::Read for UdpStream {
-  fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-    let sock = self.sock.as_ref();
-    let sock_ref = sock.deref();
-
-    sock_ref.peek_addr()
-            .validate(|rx_addr| {
-              rx_addr.should_eq(&self.addr)
-                     .else_err(|_| nb::Error::WouldBlock)
-            })
-            .bind(|_| Socket::recv(sock_ref, buf))
-            .map_err(nb_to_io)
-            .map(|Addrd(n, _)| n)
-  }
-}
-
-/// TODO
-#[derive(Debug)]
-pub struct SecureUdpSocket {
-  sock: Arc<UdpSocket>,
-  streams: Mutex<HashMap<no_std_net::SocketAddr, Arc<Mutex<UdpStream>>>>,
-}
-
-impl SecureUdpSocket {
-  /// TODO
-  pub fn new(sock: UdpSocket) -> Self {
-    Self { sock: Arc::new(sock),
-           streams: Default::default() }
-  }
-
-  /// TODO
-  pub(crate) fn get_stream(&self, addr: no_std_net::SocketAddr) -> Arc<Mutex<UdpStream>> {
-    let mut streams = self.streams.lock().unwrap();
-    let stream_ent = streams.entry(addr);
-
-    stream_ent.or_insert(Arc::new(Mutex::new(UdpStream::new(self.sock.clone(), addr))))
-              .clone()
-  }
-}
-
-impl Socket for UdpSocket {
-  type Error = io::Error;
-
-  fn send(&self, msg: Addrd<&[u8]>) -> nb::Result<(), Self::Error> {
-    self.set_nonblocking(true)
-        .bind(|_| {
-          UdpSocket::send_to::<std::net::SocketAddr>(self,
-                                                     msg.data(),
-                                                     addr::no_std::SockAddr(msg.addr()).into())
-        })
-        .map(|_| ())
-        .map_err(io_to_nb)
-  }
-
-  fn recv(&self, buffer: &mut [u8]) -> nb::Result<Addrd<usize>, Self::Error> {
-    self.set_nonblocking(true).unwrap();
-    self.recv_from(buffer)
-        .map(|(n, addr)| Addrd(n, addr::std::SockAddr(addr).into()))
-        .map_err(io_to_nb)
-  }
-
-  fn bind_raw<A: no_std_net::ToSocketAddrs>(addr: A) -> Result<Self, Self::Error> {
-    let addrs = addr.to_socket_addrs()
-                    .unwrap()
-                    .map(|no_std| addr::no_std::SockAddr(no_std).into())
-                    .collect::<Vec<std::net::SocketAddr>>();
-
-    UdpSocket::bind(addrs.as_slice()).perform(|s| s.set_nonblocking(true).unwrap())
-  }
-
-  fn join_multicast(&self, addr: no_std_net::IpAddr) -> Result<(), Self::Error> {
-    match addr::std::Ip::from(addr::no_std::Ip(addr)).0 {
-      | std::net::IpAddr::V4(addr) => {
-        self.join_multicast_v4(&addr, &std::net::Ipv4Addr::UNSPECIFIED)
-      },
-      | std::net::IpAddr::V6(addr) => self.join_multicast_v6(&addr, 0),
-    }
-  }
-
-  fn peek(&self, buffer: &mut [u8]) -> nb::Result<Addrd<usize>, Self::Error> {
-    std::net::UdpSocket::peek_from(&self, buffer).map(|(n, addr)| Addrd(n, addr::no_std::SockAddr::from(addr::std::SockAddr(addr)).0)).map_err(io_to_nb)
-  }
-}
-
-mod addr {
-  pub(crate) mod std {
+pub(crate) mod std {
     use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6};
 
     use super::no_std;
@@ -281,18 +173,3 @@ mod addr {
       }
     }
   }
-}
-
-fn io_to_nb(err: io::Error) -> nb::Error<io::Error> {
-  match err.kind() {
-    | io::ErrorKind::WouldBlock => nb::Error::WouldBlock,
-    | _ => nb::Error::Other(err),
-  }
-}
-
-fn nb_to_io(err: nb::Error<io::Error>) -> io::Error {
-  match err {
-    | nb::Error::WouldBlock => io::Error::from(io::ErrorKind::WouldBlock),
-    | nb::Error::Other(err) => err,
-  }
-}
